requirements.txt
# Core FastAPI & Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pyjwt==2.8.0
bcrypt==4.0.1

# HTTP & Networking
httpx==0.25.2
requests==2.31.0
aiofiles==23.2.1
aiohttp==3.9.1

# Docker & System
docker==6.1.3
psutil==5.9.6

# Configuration & Validation
jsonschema==4.20.0
pyyaml==6.0.1
jinja2==3.1.2
python-dotenv==1.0.0

# Validation & Utilities
validators==0.22.0
email-validator==2.1.0
click==8.1.7
humanize==4.9.0
croniter==1.4.1

# Monitoring & Metrics
prometheus-client==0.19.0

# Logging
python-json-logger==2.0.7
structlog==23.2.0

# Caching & Background Tasks (optional)
redis==5.0.1
celery==5.3.4

# Email & Templating (optional)
aiosmtplib==3.0.1

# Security dependencies  
user-agents==2.2.0
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.0.1

# Production server
gunicorn==21.2.0
